refactor(app): extract the connection-form wiring
First slice of breaking up `fn main`. `AppState` bundles the shared
`Rc`/`Arc` state so a handler no longer opens with a stack of individual
`let x = x.clone();` lines naming exactly what it captures — which is
what kept every callback pinned inside `main`. State declared partway
down `main` moves up next to the rest so the bundle can be built in one
place.

The connection form's handlers (add, add-in-group, edit, group/engine
changed, import URL, cancel, export, quick test, test, save, delete) move
to `wire::conn_form` unchanged; they destructure what they need out of
`AppState` at the top.

`AppState` carries only the fields the moved handlers use; the rest join
as further clusters come out.
